Implemented home.tsx / Prelim addition of Popup
Added the Homepage where all threads will be displayed from, adjusted the Root and Index.css to accommodate the changes.
Also added a WIP version of ThreadPopup - currently does not work and will require the use of state communication between components, or a rewrite of ThreadUI.

Finally, adjusted the implementation of thread dates, converting from Date objects to string.
This ensures ease of translation from JSON and the object Thread.