Added New Thread, Logout, and Comment tasks.
Moved isHandlingPost to the persistent slice. Added Logout, New Thread functionality.

Added Delete and Edit for comments.
Delete and Edit are not planned for threads - they feel counterintuitive to a gossip style thread where anything goes, and adding it to comments is a proof of concept.

Note: Currently considering compressing the backend URLs together so it fits the API model.